优化数据处理逻辑，修复描述字段换行问题，调整数据类型，增强错误处理，更新测试用例以支持新逻辑

+++ src/java/main/Database/InfoItem/InfoTor/InfoTorFetch.java
@@ -6,7 +6,7 @@ public class InfoTorFetch extends InfoTor {
 
     public Integer ANI_ID;
     public OffsetDateTime air_datetime;
-    public Integer size;
+    public Long size;
     public String url_page;
     public String title;
     public String subtitle_group;
